Fix incorrect declaration of randombytes() in tweetnacl.c

+++ tweetnacl.c
@@ -7,7 +7,7 @@ typedef unsigned long u32;
 typedef unsigned long long u64;
 typedef long long i64;
 typedef i64 gf[16];
-extern void randombytes(u8 *,u64);
+extern void randombytes(u8 *, unsigned int);
 
 static const u8
   _0[16],
